src/requirements.txt
aiogram>=3.0
aiohttp~=3.9.1
apscheduler
fastapi>=0.109.1
numpy >= 1.2
httpx
pillow >= 9.0
python-multipart
pyyaml~=6.0.1
torch
uvicorn
opencv-python
insightface==0.7.3
yoomoney==0.1.0
sqlalchemy
googletrans==4.0.0-rc1